VOIP-7106: ASoC: cygnus: refactor how driver accesses io
Create a wrapper around the driver's io acesses. This should
result in no functional change to the driver.
The new access api uses an "update bits" mask which makes it much
more explicit about which bits are being update. This style is
desirable because it is the same style used by the regmap api
which will be the eventual underlaying access method for this driver.
There is also a change to how the registers are enumerated.
A tag has been added to the upper bits. This tag is used to identify
which of the multiple io spaces the register belongs to.
Also the register offset is written as it appears in the datasheet,
not the offset into its io space. The wrappers will do the math
to create the correct io offset. This makes it a bit easier to
match the register definitions in the code with the datasheet.

Change-Id: Ie2d5c0af8d9d0470653d9a701e583ce328861965
Reviewed-on: http://gerrit-ccxsw.broadcom.net/49402
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: CCXSW <ccxsw.build@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: CCXSW <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>