VOIP-7106: ASoC: cygnus: Convert io access to regmap api
Use the regmap api to access io space.  This change is not being done
to fix any problems.  Upcoming changes will require the use of the regmap
api due to the need to use syscon for some of the register access.
Accessing all the register using the same api will keep the driver
consistent.  Also, the regmap interface allows for some useful debugging
features so it seems like a better way to do io accesses.

Change-Id: Iff6f292ae2b5c048c2093cf3db771ac167297b59
Reviewed-on: http://gerrit-ccxsw.broadcom.net/52511
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: CCXSW <ccxsw.build@broadcom.com>
Tested-by: CCXSW <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>