SOC-6462: Set 40-bit DMA mask for entire SoC
When applying dma-ranges property to any device, the parent node is looked
up with the property. No dma-ranges is defined in the parents of DMA
capable devices in PCIe RC, SATA, USB, FS4 and HSLS. Hence the default gets
applied to the devices.

Specify SoC wide 40-bit DMA mask to define the limits.

Change-Id: I19db38d247c26e4dccd3f14981126a5d18499e3f
Reviewed-on: http://gerrit-ccxsw.broadcom.net/53136
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: CCXSW <ccxsw.build@broadcom.com>
Tested-by: CCXSW <ccxsw.build@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>