SOC-6463: SR+ SATA changes
Register address changes are done for SR+
In SR Ax chips 8 selected SATA ports are chosen from 20 ports
which does not have issues of Linkup in Gen3.
Those 8 are "9, 13, 6, 7, 15, 16, 17, 18". So SATA DTS
file has ahci and phy nodes specific to these selected ports.
In SR+ HW, not used 12 ports are removed and used ports
are mapped to starting register addresses.
So These changes are reflecting the address changes.

Now SR+ changes are overwritten on SR Ax chip changes,
because SATA will not be used in SR Ax chip.

Change-Id: I7efc3f3216fd1c2e5eed29ec57693082491e2b4d
Signed-off-by: Srinath Mannam <srinath.mannam@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/53787
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>