SOC-6039: PCI: iproc: handle failures during probe
Current code does not take care of releasing taken resources
in case of failure(s).

For example, timer setup is done before inbound/outbound mapping
configuration and if one of the mapping fails, timer is still running,
but the pcie resources including timer callback function are freed,
resulting in crash when timer expires for the first time.

This issue is seen during kexec reboot. It is hidden in primary
kernel because mapping does not fail there, but in secondary
kernel inbound mapping fails.

Change-Id: Ib83b01e81bac70206ec4b9a90b00ad5ab5ab322b
Signed-off-by: Abhishek Shah <abhishek.shah@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/53842
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: CCXSW <ccxsw.build@broadcom.com>
Reviewed-by: CCXSW <ccxsw.build@broadcom.com>
Reviewed-by: Vikram Prakash <vikram.prakash@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>