VOIP-7327 pinctrl: iproc: Fix GPIO pad override logic
On Cygnus, DTE block does not have hysteresis pad configuration and
causes the delta between time stamps to become unstable.

Added field "ctrl_path_override" to pin function group. Setting it to 1
allows GPIO pad configuration to be exposed to external pad, even for
pins that are not muxed as GPIO. Therefore for pins that are muxed
DTE, the input hysteresis can still be turned on through GPIO pad.

Currently the override bit is only allowed for DTE pins.

Change-Id: If548b34ce12e878953084cd2f69c064df635483f
Reviewed-on: http://gerrit-ccxsw.broadcom.net/56515
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Li Jin <li.jin@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>