VOIP-7106: ASoC: cygnus: Fix minor issue with regmap usage
Fix a small error in calculating the the max_register for the cfg
of regmap space. The max_register value was too big.  This would not
result in any real problem, except if the user was going to make
an out of bounds access.
Now we cacluate the max register value from the devicetree "reg"
propertry which is already specifying devices the io range.

Change-Id: I2cde34f9b32e800d9ebc327d3d81e85799a665a1
Reviewed-on: http://gerrit-ccxsw.broadcom.net/57257
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>