VOIP-7454: ASoC: cygnus: Convert driver to use audio mux
Convert the audio driver to use the audio mux which is now accessed
through the clock framework. Previously, the audio driver accessed
this mux directly. There should not be any functional change to the
system, but now we can use standard clock binding in the the device
tree when configuring the clock.

Change-Id: I17d77363cd976f1015dece4a90630416f7458500
Reviewed-on: http://gerrit-ccxsw.broadcom.net/59004
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>