VOIP-7468: cygnus: dts: Fix audio mux clock parent assignment
When using the standard assigned-clocks binding we can not have
multiple entries in the same node, only the first instance is
used. The symptoms of this mistake would show up if using
two ports at the same time that required different clock settings.
The second port to start would override the first.

Change-Id: Ibb045eaee155252be366a0d255be649bbe8bcb4c
Reviewed-on: http://gerrit-ccxsw.broadcom.net/60063
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Lori Hikichi <lori.hikichi@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>