VOIP-7479: clk: iproc: Make clock-output-names optional for iproc pll
The iproc pll driver used to require clock-output-names be specified.
Now that is optional, if the property is not specified, then the name
of the pll leaf clocks will be automatically generated by the driver.
The name will be the name of the parent pll appended with _chX, where
X is the numerical index of the leaf clock (e.g., audiopll_ch0).

Change-Id: Ic7aa77e3c5960ad693b59a9eeaeb075e5abe98c5
Reviewed-on: http://gerrit-ccxsw.broadcom.net/60350
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>