VOIP-7356: omega: add cpuidle c-states
entry and exit latency values are derived by profiling low
level sequences on m7. PREP time is taken form standard atf
configuration to be ~20us for core and ~60uS for cluster.
based on that wakeup latency is taken as
 entry latency + exit latency - PREP

min residency values needs to be tuned for power optimization.

with this configuration impact of cpuidle on rest of the
system can be tested and issues, if any can be discovered.

NOTE: crmu CONFIG_DEBUG and CONFIG_PRINT needs to be off
as it affects latencies significantly inturn will affect
scheduling.

Change-Id: I4eb0af8041df6657f6c8eb821947570b9c03d128
Signed-off-by: Sandeep Tripathy <sandeep.tripathy@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/60855
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
(cherry picked from commit edc698d971f6f933d41bad62f1adf2186bdfac21)
Reviewed-on: http://gerrit-ccxsw.broadcom.net/68219
Reviewed-by: Vladimir Dreizin <vladimir.dreizin@broadcom.com>
Tested-by: Vladimir Dreizin <vladimir.dreizin@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>