VOIP-7434: net: phy: Modified register offsets in Broadcom iProc mdio mux
Modified the register offsets in the Broadcom iProc mdio mux to start
from the top of the register address space.

Earlier the base address specified was from the middle of the block's
register space. The new base address will be of the start of the mdio's
register address space. The offsets have been fixed to match this..

Fixes: 98bc865a1ec8 ("net: mdio-mux: Add MDIO mux driver for iProc SoCs")

Change-Id: Idc4ac916dec38e59a88e4ad580ee80f7e524e6ea
Signed-off-by: Arun Parameswaran <arun.parameswaran@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/61964
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>