SOC-7851: PCI: iproc: Fix up corrupted PAXC root complex config registers
Backported change from mainline

On certain versions of Broadcom PAXC based root complexes, certain
regions of the configuration space are corrupted. As a result, it
prevents the Linux PCIe stack from traversing the linked list of the
capability registers completely and therefore the root complex is
not advertised as "PCIe capable". This prevents the correct PCIe RID
from being parsed in the kernel PCIe stack. A correct RID is required
for mapping to a stream ID from the SMMU or the device ID from the
GICv3 ITS

This patch fixes up the issue by manually populating the related
PCIe capabilities

Change-Id: I6248963271e65d510bd0e46fcff9acd15c75f067
Signed-off-by: Ray Jui <rjui@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/67484
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>