VOIP-7106: clk: iproc: update binding for audiopll 1
Audiopll 0 and 1 need to set different bits in their respective
power control registers. Therefore, the current solution was not
adequate for handling two instance of the audio pll. We need
to create a different clock compatibility string so that we can
use a different initialization table for the two plls.

Change-Id: Ica898603f8ef0b1fb5d8969f60758877f4afc473
Reviewed-on: http://gerrit-ccxsw.broadcom.net/63139
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>