VOIP-7106: clk: omega: Fix driver to enable audiopll 1
Audiopll 0 and 1 need to set different bits in their respective
power control registers. Therefore, the current solution was not
adequate for handling two instance of the audio pll. We need
to create a different clock compatibility string so that we can
use a different initialization table for the two plls.

Change-Id: I62c0e92ae908fe5c0a8c16bbdf6ddb858b9460ee
Reviewed-on: http://gerrit-ccxsw.broadcom.net/63330
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>