VOIP-7539: net: dsa: b53: Add support for Broadcom Omega internal switch
Add support for the Broadcom Omega SoC's internal switch to the b53
driver in the DSA framework.

The internal Ethernet switch in the Omega is similar to the BCM58XX
switch family, except for the BRCM tag location and the active ports.
In Omega, the BRCM tag is inserted in between the mac address and ether
type and  only the ports 0,1 and 8(cpu port) are active.

So, added a new switch device (BCM583XX) to the B53, under the BCM58XX
family, to handle the differences in Omega.

Change-Id: I5efb8fa0ceb7835e3e6ec265281d767f44e0f9f2
Signed-off-by: Arun Parameswaran <arun.parameswaran@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/63877
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>
Tested-by: Scott Branden <scott.branden@broadcom.com>