VOIP-7580: mmc: sdhci-iproc: Add SDHCI_QUIRK_CAP_CLOCK_BASE_BROKEN
The sdhci iproc driver makes it manadatory during driver probe that a
clock is provided for the purpose of querying the rate. This is the
intended method of communicating to the framework the max clock rate.

This has not been a problem to date because the max clock field of the
capabilities register has been set to 0. In this case the driver will
query the provided clock for the max rate. If it is the case that the
capability is set to non-zero, then the provided clock will not
be queried and could result in mismatched clock rates.

Setting SDHCI_QUIRK_CAP_CLOCK_BASE_BROKEN will force the framework
to always query the provided clock.

Change-Id: I20144aae2fc8fcb30994a97af2ed831878ff5f5a
Reviewed-on: http://gerrit-ccxsw.broadcom.net/65071
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>