SOC-7481 misc: bcm-vk: add additional error codes for read/write functions
Currently, the only error condition for the read/write is to return
0 as the ssize.  FFMPEG needs a bit more refined returned error codes
which are added.  For reading, it will return -ENOMSG when queue is
empty, and -EMSGSIZE if the buffer length is not big enough to store
the first response.  For writing, default error is -EPERM, and it will
return -EOVERFLOW if the message length is bigger than the ring buffer's
max tolerable, -EADR for bad request if the size is not a multiple
of the basic block size, and -NOMEM for malloc issue.

Change-Id: Ie65261a8d89184544ddccd6d4b6cf59c891fd0c8
Reviewed-on: http://gerrit-ccxsw.broadcom.net/67849
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>