SOC-7481: misc: bcm-vk: add msgq init state to allow more flexible probing
When the driver is probed, there are valid cases that the msgq ring
descriptors may not have been initialized by the VK-OS.  A state is
introduced in the driver so that it could be inserted without msgq init.
Some of the functions such as read/write or interrupt handling will be
skipped if msgq is not initialized.
If the initialized marker is detected, however, the logic will work as
before without any behavioral changes.

Change-Id: Id93c225d5f35db6300a7ed945a5917e7f3616cef
Reviewed-on: http://gerrit-ccxsw.broadcom.net/70584
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>