VOIP-7106: ASoC: omega-wa: Add control of ak4458 reset
Add control for gpio reset of ak4458. Also, add in pinmux control
for the mfio 45 and 46 pins. There is a problem where the ak448
codec glitches the SCL line during reset. This causes Omega's i2c
controller to think that another master has taken conrol of the
bus (i.e, the bus is busy). In order to avoid this issue we
flip the pins to gpio mode just prior to resetting the ak4458,
in this way the controller will not see the glitch.

Change-Id: If399c4fd90fea07aeece7064738789a92cd4ae36
Reviewed-on: http://gerrit-ccxsw.broadcom.net/71097
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>