SOC-8184: mailbox: bcm-flexrm-mailbox: Schedule irq thread only when data is available
During high data transfer peek_data() function can be called separately by
dma driver (bcm-sba-dma) when there is are no free request. Hence it is
possible that peek_data() process all data and there is nothing to letf
when interrupt is generated.

To avoid stacking of irq thread it is safe to schedule irq thread only if
there is data to process.

Change-Id: I8ae3a6e1735a4f66d51650bafac06f6357f5e1da
Signed-off-by: Rayagonda Kokatanur <rayagonda.kokatanur@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/72156
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>