SOC-8407: phy: Fix system hang from the SR USB PHY driver
The current Stingray USB PHY driver does not handle the return value of
several PHY initialization routines. As a result, even when PHY init has
failed (e.g., USB PHY PLL failed to lock), the driver will still return
zero as success to the PHY framework. This will result in the client
driver (e.g., BDC or USB host) attempts to proceed by accessing some of
the registers when the PLL is not even up and therefore causes a bus hang.
The fix is to handle the return value of these PHY init routines properly
and make sure they are passed up to the PHY framework

Change-Id: Ie22f48ccaf996fc8cf77ec2b7cd4ae4e4a26f5a2
Fixes: 1b1cebd236a7 ("phy: bcm-stingray-usb: initialize phy usb driver in generic manner")
Signed-off-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/74581
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>