SOC-8058: phy: bcm-sr-pcie-diag: Rename pmi read/write functions
Rename the following functions: 'pmi_write_via_paxb' and
'pmi_read_via_paxb' to 'pmi_write' and 'pmi_read'.

The access to the PCIe serdes via PMI when in root complex mode is
always done by PAXB on Stingray; therefore, the 'via_paxb' phrase is
redundant and can be removed.

Change-Id: I810257762460b5b221f336d5b45574484f7b1857
Signed-off-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/75116
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>