SOC-8058: phy: bcm-sr-pcie-diag: Fix PCIe lane programming
In unicast mode, the PMI address requires the lane offset to be
specified as:
UC0 - for lane 0 to 3
UC1 - for lane 4 to 7
UC2 - for lane 8 to 11
UC3 - for lane 12 to 15

The actual lane number to be used depends on the above lane offset.

The current PMI implementation does not program the lane offset; as a
result, the PMI interface can only communicate to lanes 0, 1, 2, 3 and
communications to lanes 4 - 15 will not work. While this is not an issue
on Stingray since each PCIe serdes has only two lanes, the
implementation should be improved to reflect the actual bit fields of the
PMI interface.

Improve the driver by properly programming the lane offset field and
using a new imeplementation to derive the PMI address.

Change-Id: Ie6b9af4e1b51078e969ce3243b87e5cfb791c204
Signed-off-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/75157
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Abhishek Rajeshkumar Shah <abhishek.shah@broadcom.com>