VOIP-7563: misc: dte: Add base Omega support to the driver
Add support for the DTE block on Omega.
As usual, the client mapping is different for Omega, but the other
major change is that the io space mapping has changed.
Now instead of constant defines for register offset we will store
the information for each SoC in the parameteric configuration table.
Also, create a dte io access layer so we can hide the details about
which io block a specific reg resides in.
This should result in no functional change for the previous chips.

Change-Id: I975ca159e4ee7052371be4982692700b25c87d09
Reviewed-on: http://gerrit-ccxsw.broadcom.net/76055
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>