VOIP-7784: ASoC: audioh: Power down EP interface when not in use
Setup the EP path so that it is controlled by DAPM. This will
allow the interface to only be powered when in use. The handset
and headset controls were adjusted to create a virtual demux. The
demux will enforce the rule where only one of the handset or
headset should be powered at the same time.

Change-Id: I7f0854d2bdc8934cb7e8036395e263e0821391ff
Reviewed-on: http://gerrit-ccxsw.broadcom.net/78719
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>