SOC-8792: misc: bcm-vk: Add readback, verify and retry in h2vk communication
The FPGA board seems to have some timing issues where occasionally,
data written from host to vk may show up as "corrupted" value (even when
reading it back right the way).  This workaround will perform a readback
and verify, and if the values are different, repeat the same write until
values do match up.

It is aimed as a temporary workaround for vk2h rd_idx corruption
(kernel panic), or data not landing to vk properly (unsupported message
+ corrupted buffer ptr) issues. Note that HW/ASIC has been informed and
when the issue is fixed, this enhancement could be turned off.

Change-Id: I40b7dd33611fc1c3b0be76135f0b13538ef065c4
Reviewed-on: http://gerrit-ccxsw.broadcom.net/78822
Reviewed-by: Scott Branden <scott.branden@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Shamez Kurji <shamez.kurji@broadcom.com>