SOC-5418 drivers: xhci: Add quirk to do USB port PHY reset
Stingray USB HS PHY has an issue, that USB HS device is detected
as FS device after we use FS device in the same port. This problem
can be resolved after reset the HS phy when the device is disconnected.
Add a quirk to do USB port PHY reset when port disconnect event.

Change-Id: Ic25355a7bc8b2dc748d725f53bb2b2a9a8e312ec
Signed-off-by: Srinath Mannam <srinath.mannam@broadcom.com>
Reviewed-on: https://gerrit-ccxsw.broadcom.net/139829
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>