SOC-5418: drivers: xhci: Add quirk to do USB port PHY reset
Stingray USB HS PHY has an issue, that USB HS device is detected
as FS device after we use FS device in the same port. This problem
can be resolved after reset the HS phy when the device is disconnected.
Add a quirk to do USB port PHY reset when port disconnect event.

Change-Id: Ic7005ad0db5b7768bd5844c1e2ccd5f4bac268cf
Signed-off-by: Srinath Mannam <srinath.mannam@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/79026
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>