VOIP-8039: arm64: dts: bcm958313-wa: Set boot-on property of vdd regulator
Set the regulator-boot-on property for the vdd regulator.
We want to inform regulator framework that this regulator is
already on. Without this property, during regulator initialization,
the framework would enable the regulator and turn it off briefly before
turning in on.  This momentary change is state would sometime cause our
vdd rail to dip, causing boot problems. Now the regulator will stay on
during the initialization stage.

Change-Id: I95bfa08e313f24b35d29ff7624d6f8736ee3acd9
Reviewed-on: http://gerrit-ccxsw.broadcom.net/79091
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Vladimir Dreizin <vladimir.dreizin@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>