VOIP-8029: staging: brcmesw: Use fixed size types for esw_reg_data
Change the 'len' field from size_t to a fixed size uint8_t.
size_t will vary in size depending on the compiler.
Since this file is shared with user space if we build user
space with 32 bit compiler and the kernel is 64 bit then
we get mismatched structures. This causes problems when
communicating the data to the kernel.

Change-Id: I3c62ff3d80970bdf4cc25eb897163bed969473c4
Reviewed-on: http://gerrit-ccxsw.broadcom.net/79537
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Arun Parameswaran <arun.parameswaran@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>
Tested-by: Scott Branden <scott.branden@broadcom.com>