Revert "VOIP-7873: net: dsa: Fix to store new mtu setting in netdevice."
Default approach used by all existing network drivers is to store
modified mtu as part of ndo_change_mtu callback function and not
in the core driver. This approach is not optimal but it is the
accepted approach. Undoing the fix and not storing new mtu
setting in the core driver.

This reverts commit bc61859e667a9b3f6b75153533bd8646a2904545.

Change-Id: I05001fc023ae935b20278d336f509316ad12dff3
Reviewed-on: http://gerrit-ccxsw.broadcom.net/79718
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>