SOC-8775: arm64: dts: NS3: Modify SST100 PCIe IOMMU and MSI map
In SST100 platform IOMMU and MSI map is changed to
have bus width 2, device width 5 and function width 3 bits
to address 4 buses, 32 devices, and 8 functions.
It helps to connect EPs with VF feature in which device and
function numbers are more than 1.

Extending the support to NS3 form stingray using commit ID
1d661e2e91e11a112d6c728233d53aecdae26475 for PCIe_0

Change-Id: I6b4a03e78ebf406dfad0b57d60eb2c005c1fd320
Signed-off-by: Bharat Kumar Reddy Gooty <bharat.gooty@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/81260
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Srinath Mannam <srinath.mannam@broadcom.com>
Reviewed-by: Vikram Prakash <vikram.prakash@broadcom.com>