VOIP-8110: ASoC: omega-ssp: Add names to regmap configs
By default a regmap debugfs entry will be created using the driver
name as the debugfs entry.  The omega ssp driver has two regmaps
therefore, we get a name collision and the warning:
  Failed to create 600000.audio debugfs directory

Adding the regmap names allows unique debugfs entries to be created.
This problem only impacted the ability to access the debugfs entries,
so it should have no impact on normal operation.

Change-Id: I5c6bd152a10b4a311fc044f2ae28cf7d52bfa79e
Reviewed-on: http://gerrit-ccxsw.broadcom.net/82663
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Lori Hikichi <lori.hikichi@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>