VOIP-8186: ASoC: omega: Stop using regmap_read_poll_timeout
The regmap_read_poll_timeout uses a sleep() call which is not
appropriate for an atomic context. In certain situation, like when
the audio transfer needs to be restarted, this function can get called
from an atomic context. Replace the call with readl_poll_timeout_atomic.
Also, the maximum wait time has been reviewed and set to 200us to prevent
any extra long waits while polling.

Change-Id: I86d2194b24b1231adb03bde0c2505ce4d3b48118
Reviewed-on: http://gerrit-ccxsw.broadcom.net/83133
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>