SOC-8948: misc: bcm-vk: Add logic to support VK reset
Add implementation for resetting the Valkyrie M7.  When the reset
function is called, all new open to VK will be blocked while all
existing sessions will be shipped a kill signal.  At the exit of
the reset process, a card-level shutdown will be sent to VK to
trigger it to perform a soft reset.

Change-Id: I8609a3f5c4c1e7d4b4d713f5e352583bd2447597
Reviewed-on: http://gerrit-ccxsw.broadcom.net/83583
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>