SOC-9099: mmc: sdhci-iproc: Set NO_HISPD bit to fix HS50 data hold time problem
The iproc host eMMC/SD controller hold time does not meet the
specification in the HS50 mode. This problem can be mitigated
by disabling the HISPD bit; thus forcing the controller output
data to be driven on the falling clock edges rather than the
rising clock edges.

This change applies only to the Cygnus platform.

Change-Id: Ie0617fee24099a196fd77b261f4cb206f168c6bb
Fixes: aad9b776931c ("Revert "VOIP-7580: mmc: sdhci-iproc: Add SDHCI_QUIRK_CAP_CLOCK_BASE_BROKEN"")
Signed-off-by: Trac Hoang <trac.hoang@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/84592
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: JD Zheng <jiandong.zheng@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>