SOC-9252: misc: bcm-vk: Add kref to valkyrie's file operations
Valkyrie is created with PCIe + MISC driver.  When the
PCIe device is removed through its sysfs("remove"), the kernel will not
take into consideration who are still using the fops->open.  This releases
the PCIe device, and if there is still a user holding onto the fops, that
user will crash when it exits because the memory have been freed.  Add kref
to do housekeeping and free only when all are references are gone.

Change-Id: I8c5961997cb6fea72b83e974de3d6e8aab12d27e
Reviewed-on: http://gerrit-ccxsw.broadcom.net/85035
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>