SOC-9409: misc: bcm-vk: Hook up DB interrupt to vkcli reset operation
As the finalized method for soft-resetig VK is DB#3, add logic to
the driver to make it operational.
Note: There seems to be an issue (TBD) if there is FFMPEG session going
on while we bruteforcely force a reset.  The card still respond to
the DR reset, but when it runs boot1, it will be stuck at boot1 does not
go to completion.

Change-Id: Id88b0364f721320b2a1a4184aeb9c96cf12b1047
Reviewed-on: http://gerrit-ccxsw.broadcom.net/85797
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Shamez Kurji <shamez.kurji@broadcom.com>