SOC-9487 misc: bcm-vk: change bcm_vk_load_image to support boot2 up to 64M
Correct bcm_vk_load_image to load images up to 64M for boot2.
load image was writing to BAR1 previously before which happened to have a
2M window into DDR.  Use BAR2 as it has the 64M window to DDR.

Start using request_firmware_into_buf to read from file directly into
memory on BAR so that kernel does not need to malloc to RAM and then
a copy to BAR afterwards.

Change-Id: I47032bb81c35308fb614dde6cd4a8e94499c4e36
Reviewed-on: http://gerrit-ccxsw.broadcom.net/90771
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>