VOIP-8187: ASoC: bcm-audioh: ADC PGA gains not matching mixer control
The AudioH ADC PGA gains control allows for a range from 0 trough 15,
but the actual control on the chip does not use value 6 or 7.
Therefore, the effective range is from 0 to 13. To provide the user
with a sensible looking control, wrap the normal volume macros and
control fucntions with a bit of code that remaps the user's 0-13
value into a value of 0-5,8-15 that work in the register.

Change-Id: I048c97fab51d522b3ed2e29afa1befddb4b59b09
Reviewed-on: http://gerrit-ccxsw.broadcom.net/91493
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>