VOIP-8256: ASoC: bcm-audioh: Toggle reset when enabling headset
The headset interface needs to have the DAC1 reset bit toggled.
The handset and headset share the same same reset signal.
The interfaces are mutually exclusive, so it it fine to reset one
interface without worry about the other.
Also, rename the internal functions for these interfaces to highlight
how the two interface are closely related.

Change-Id: I1bb56fdd3168bc577766813e41f59c92b1c4ed6f
Reviewed-on: http://gerrit-ccxsw.broadcom.net/93049
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>