VOIP-8252: ASoC: bcm-audioh: Add workaround for io read corruption
There is a chance that the io read access to the audioH codec will
return corrupted data.  This may happen if we read a register while
actively recording from the codec.  The read has to coincide with
a frame sync.  The workaround is to read from the register multiple
times until all the values match.

Change-Id: Ic8a9567d90206992f7c50dd4dee34f74ae077423
Reviewed-on: http://gerrit-ccxsw.broadcom.net/93745
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>