SOC-9533: arm64: dts: NS3: Modify default msi and iommu map
In storage usecases stingray RC is connected to NVMe EPs through PCIe
switch in which msi and iommu map should support multiple buses and single
device and functions. So that SID is configured as 8bit bus, 1bit device
and 1bit function. And this configuration is taken as default for all SR
platforms.

Now default SID configuration is changed to 2bit bus, 5bit device and 3bit
function to support multiple function EPs connected directly to RC.

Change-Id: Id209b3acf6874b3cea889ddf9c1098bf1cfd3fa4
Signed-off-by: Srinath Mannam <srinath.mannam@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/94361
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>