SOC-9797: misc: bcm-vk: Add auto loading capability to VK
When the valkyrie card is probed by the PCIe driver, and if the
default files exist in the default directory, perform auto
loading of the required binaries to bring it to running state.
This feature will be controlled by a module parameter "auto_load"
which by default is true, ie, enabled.

Change-Id: I2c9cfa9a3a33161b80e149a6efc92240d7c552b8
Reviewed-on: http://gerrit-ccxsw.broadcom.net/95315
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Shamez Kurji <shamez.kurji@broadcom.com>