SOC-9565: LPM Backup: Update Live backup state in SSR
A new field “live_backup_state” is added to SSR to indicate Live backup
state.  It is not updated at each live backup stage, anticipating for
lesser time for LPM switch in case LPM interrupt is received.

This field will be initialized when ARM command is issued and
gets updated by result at the end of live backup.

If we switch to LPM mode during Live backup, CRMU updates the
field while in LPM.

Change-Id: I06d965d8bc2426faa127922a7479a6e1eea6d0d4
Signed-off-by: Abhishek Shah <abhishek.shah@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/95245
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>