SOC-10010: misc: bcm-vk: Add mutex for boot1 or boot2 download
With the autoload feature enabled, there will be xN parallel
downloads, where N is the number of inserted valkyries.  Experiments
seem to indicate that the firmware request API do not seem to allow
this, so adding a mutex for protection.  Will need to re-visit the
lib API to see why mal-function happens.

Change-Id: Ie98bc416a074d8254fee7dd8624a36a3ee1ed2b1
Reviewed-on: http://gerrit-ccxsw.broadcom.net/96132
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Tested-by: Suji Velupillai <suji.velupillai@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: Jeffery Li <jeffery.li@broadcom.com>
Reviewed-by: Shamez Kurji <shamez.kurji@broadcom.com>