SOC-10231: misc: bcm-vk: Append SGL to end of message's reserved buf
Add a module parameter which will make the driver create a message
that has extra room for short SGL storing.  The extra buffer will
be filled with short SGLs if possible, so that the receiving end do
not need to perform an extra DMA to get the SGL first.  This is one
of the performance improvement initiative.

Change-Id: I071de9092ef6dafbd71592671feeed6da186abe5
Reviewed-on: http://gerrit-ccxsw.broadcom.net/98802
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>